Add alpha shower

+++ app/src/main/java/com/neurotechx/smartphonebci/driver/dsp/SlidingWindow.java
@@ -27,7 +27,7 @@ public class SlidingWindow {
     }
 
 
-    public void push(short data) {
+    public void push(float data) {
         mQueue.add((double) data);
         mWindowStep++;
         if (!mQueue.isAtFullCapacity()){
@@ -42,8 +42,8 @@ public class SlidingWindow {
         }
     }
 
-    public void push(short data[]) {
-        for (short d : data) {
+    public void push(float data[]) {
+        for (float d : data) {
             this.push(d);
         }
     }
